+++ Farmera/src/utils/AxiosConfig.jsx
@@ -18,10 +18,21 @@ const SetUpAxiosInterceptors = () => {
     );
 
     // Response interceptor
+    // axios.interceptors.response.use(
+    //     (response) => response,
+    //     (error) => {
+    //         if (error.response?.status === 401) {
+    //             localStorage.clear();
+    //             window.location.href = '/signin';
+    //         }
+    //         return Promise.reject(error);
+    //     }
+    // );
+
     axios.interceptors.response.use(
         (response) => response,
         (error) => {
-            if (error.response?.status === 401) {
+            if (error.response?.status === 401 && window.location.pathname !== '/signin') {
                 localStorage.clear();
                 window.location.href = '/signin';
             }
